chore: createShadowのdefaultShadowの定義元を分ける
